requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
gcsfs
matplotlib
tensorflow
joblib
scikit-learn
comet-ml
dvc